ms_proxies/requirements.txt
aiohttp==3.11.10
fastapi==0.112.2
motor==3.6.0
pymongo==4.9.2
python-dotenv
aiohttp_socks
aiohttp[speedups]
requests
stem
uvicorn==0.30.6
